Move general User and Client model to proper model files
